requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-core==1.29.4
backoff==2.2.1
beautifulsoup4==4.12.2
boilerpy3==1.0.6
canals==0.8.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
docopt==0.6.2
Events==0.5
exceptiongroup==1.1.3
farm-haystack==1.21.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
inflect==7.0.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langdetect==1.0.9
lazy-imports==0.3.1
llvmlite==0.41.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numba==0.58.0
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
openai-whisper==20230918
packaging==23.2
pandas==2.1.1
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
Pillow==9.0.0
platformdirs==3.11.0
posthog==3.0.2
prompthub-py==4.0.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pypdf==3.16.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-frontmatter==1.0.0
python-magic-bin==0.4.14
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sseclient-py==1.8.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tika==2.6.0
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.8.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.6
yarl==1.9.2
